supersu installation, move instead of unzip
The supersu files are all now uncompressed in tools/supersu, so now we don't need to uncompress them before the installation.
The supersu files are extracted at the beginning of the script, but not in the recommanded path for installation, so i move the files with a mv command.
After that i start the installation with the usual way.