regroup some aroma files
The aroma stay sensibly the same between each devices, so i grouped the files in one folder.
The common are the files will be used on all devices, no dependencies on specific architectures or hardware.
The 1080p folder are fo the 1080p devices, it contains the splash screen and the theme of aroma installer.
Obviously the "oneplus3" folder is for the OnePlus3 device, it contains the files for elementalx kernel.