kernel: patch: patch kernel before supersu
Patching the kernel after the supersu can cause issue. Supersu perform a backup of the kernel ramsdik before the kernel patch and save it in the data partition.
Flashing supersu over the installation will restore the unpatched kernel and will cause issue.
Now the kernel is patched before the supersu installation, and will remove the dm-verity and forceencryption.
Else supersu will do that as it shoudl.