aroma: switch to form for system apps
It will make the aroma more andy to use, i will no more use the "item.1.9" but the tag given to the specific entry.
For exemple if an option in removed in the aroma, curreently we need to update the updater script witht the new value. Now with the tag, it's possbile to remove any option witout any issue. The current system app remover is no more funtional, a new one will handle the form with more versatility.