Hi Nevax,
Update Turkish language.

When I installed the latest version, I noticed that on the install screen, the average of 30% of the translations I have made. Almost all in English.
When I check it, there seems to be a part here, I did the translation in Turkish. But the English part of the installation screen is dialed.

There are also many sections that need to be added to this file. It would be much nicer to add in the missing parts.

Select the Turkish language on the installation screen and you will understand the problem I am talking about. As if the translation was made halfway.

Have a nice day

Mevlüt